Main: move ShadowRenderer implementation to own file
reducing SceneManager LOC by 1/3
